bugfix: pytables silently ignores empty dataframes when writing to h5, added a row of nan to get around this

+++ single_cell/workflows/hmmcopy/__init__.py
@@ -104,7 +104,7 @@ def create_hmmcopy_workflow(
             'hmmcopy/segments'
         ),
         kwargs={
-            'dtypes': {'end': int, 'median': float}
+            'dtypes': {'end': float, 'median': float}
         }
     )
